MessageBox added before deletion

+++ RecordsManagementClient/Components/AllRecords.xaml.cs
@@ -58,41 +58,49 @@ namespace RecordsManagementClient.Components
 
         private void btnDeleteRecord_Click(object sender, RoutedEventArgs e)
         {
-            if(ManegementWindow.currentAdmin == null)
+            if (ManegementWindow.currentAdmin == null)
             {
                 MessageBox.Show("There is no admin currently logged in!");
                 return;
             }
 
-
             if ((dataGrid.SelectedItem as Record) != null)
             {
-                RestRequest request = new RestRequest(ManegementWindow.RestURL, Method.Delete);
-                Dictionary<string, object> jsonObject = new Dictionary<string, object>();
-                jsonObject.Add("current_admin_name", ManegementWindow.currentAdmin!.AdminName);
-                jsonObject.Add("current_admin_password", ManegementWindow.currentAdmin!.AdminPass);
-                jsonObject.Add("recordid_to_delete", (dataGrid.SelectedItem as Record)!.Id);
-
-                var json = JsonSerializer.Serialize(jsonObject, typeof(Dictionary<string, object>));
-                request.AddBody(json);
-
-                var response = ManegementWindow.Client.Delete(request);
-
-                if (response.StatusCode != System.Net.HttpStatusCode.OK)
-                    MessageBox.Show(response.StatusDescription);
-                else
+                var result = MessageBox.Show($"Are you sure to delete {(dataGrid.SelectedItem as Record)!.Performer} - {(dataGrid.SelectedItem as Record)!.Title}?",
+                                    "DELETE", MessageBoxButton.YesNo);
+                if (result == MessageBoxResult.Yes)
                 {
-                    Response responseFromDelete = ManegementWindow.Client.Deserialize<Response>(response).Data!;
-                    if (responseFromDelete.Error == 0 && responseFromDelete.Message == "Deleted successfully!")
-                    {
-                        MessageBox.Show(responseFromDelete.Message);
-                        RefreshRecordsGrid();
-                    }
+
+
+                    RestRequest request = new RestRequest(ManegementWindow.RestURL, Method.Delete);
+                    Dictionary<string, object> jsonObject = new Dictionary<string, object>();
+                    jsonObject.Add("current_admin_name", ManegementWindow.currentAdmin!.AdminName);
+                    jsonObject.Add("current_admin_password", ManegementWindow.currentAdmin!.AdminPass);
+                    jsonObject.Add("recordid_to_delete", (dataGrid.SelectedItem as Record)!.Id);
+
+                    var json = JsonSerializer.Serialize(jsonObject, typeof(Dictionary<string, object>));
+                    request.AddBody(json);
+
+                    var response = ManegementWindow.Client.Delete(request);
+
+                    if (response.StatusCode != System.Net.HttpStatusCode.OK)
+                        MessageBox.Show(response.StatusDescription);
                     else
                     {
-                        MessageBox.Show(responseFromDelete.Message);
+                        Response responseFromDelete = ManegementWindow.Client.Deserialize<Response>(response).Data!;
+                        if (responseFromDelete.Error == 0 && responseFromDelete.Message == "Deleted successfully!")
+                        {
+                            MessageBox.Show(responseFromDelete.Message);
+                            RefreshRecordsGrid();
+                        }
+                        else
+                        {
+                            MessageBox.Show(responseFromDelete.Message);
+                        }
                     }
                 }
+                else
+                    return;
             }
             else
                 MessageBox.Show("There is no record selected!");
